rename traces to trace
