+++ opentelemetry-user-events-trace/src/trace/mod.rs
